Initial First Attempt
Created the JavaScript that would grab the data from all three different time zones and displays them in a 24 hour format. HH:MM. I also included a SVG that would update depending on the time zone, when it's morning it will become white, however when it's dark it will be black indicating night.

Created using HTML, CSS and JavaScript.